✨ 유류집계 등록 API 추가

+++ src/main/java/com/lineinc/erp/api/server/domain/fuelaggregation/entity/FuelInfo.java
@@ -85,7 +85,7 @@ public class FuelInfo extends BaseEntity {
      */
     @DiffInclude
     @Column
-    private Double fuelAmount;
+    private Long fuelAmount;
 
     /**
      * 비고 (메모)
